+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG for FlatCAM beta
 28.12.2020
 
 - Gerber Editor - Import Shape sub-tool finished (by adding selection by dragging a selection box)
+- fixed issue #478 - Tcl Command 'New' updated
 
 27.12.2020